+++ automad/src/client/admin/components/System/CacheMonitor.ts
@@ -53,13 +53,17 @@ class CacheMonitorComponent extends BaseComponent {
 			cacheMonitorDelay.value = App.system.cache.monitorDelay;
 		});
 
+		const options = [60, 120, 300, 600];
+
+		if (!options.includes(App.system.cache.monitorDelay)) {
+			options.push(App.system.cache.monitorDelay);
+		}
+
 		createSelect(
-			[
-				{ value: 60, text: '1 min' },
-				{ value: 120, text: '2 min' },
-				{ value: 300, text: '5 min' },
-				{ value: 600, text: '10 min' },
-			],
+			options.map((o) => ({
+				value: o,
+				text: `${Math.round(o / 60)} min`,
+			})),
 			`${App.system.cache.monitorDelay}`,
 			this,
 			'cacheMonitorDelay',
@@ -69,7 +73,8 @@ class CacheMonitorComponent extends BaseComponent {
 			{
 				[Attr.bind]: 'cacheMonitorDelay',
 				[Attr.bindTo]: 'value',
-			}
+			},
+			'AM_CACHE_MONITOR_DELAY'
 		);
 	}
 }